Add integration tests for wallet and transaction repos
In-memory SQLite tests for: wallet CRUD, balance update with optimistic
lock verification, duplicate detection; transaction CRUD, soft-delete,
time-range filtering, category upsert idempotency.

Fix Wallet.Version type int -> int64 (styx SQLite scanner returns int64
from raw SQL, causing reflect panic on re-read after UPDATE).

Note: styx InsertOne skips zero-value fields, leaving deleted_at/version
as NULL. Tests work around this with non-zero Version seeds and explicit
NULL->0 fixup for deleted_at.

+++ models/balance.go
@@ -21,7 +21,7 @@ type Wallet struct {
 	Balance          float64
 	LastTxnAmount    float64
 	LastTxnTimestamp int64
-	Version          int `db:"version"`
+	Version          int64 `db:"version"`
 }
 
 type Event struct {
